+++ misk/web/@misk/common/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/common",
-  "version": "0.0.28",
+  "version": "0.0.29",
   "description": "Microservice Kontainer Common Libraries, Externals, Styles",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
   "main": "lib/common.js",
@@ -50,7 +50,7 @@
     "styled-components": "^3.4.2"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.15",
-    "@misk/webpack": "^0.0.16"
+    "@misk/dev": "^0.0.20",
+    "@misk/tslint": "^0.0.3"
   }
 }
