+++ misk/web/@misk/common/tslint.json
@@ -1,3 +1,3 @@
 {
-  "extends": "@misk/dev"
+  "extends": "@misk/tslint"
 }
