+++ misk/web/@misk/components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@misk/components",
-  "version": "0.0.18",
+  "version": "0.0.19",
   "description": "Microservice Kontainer Common Components",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
   "main": "lib/components.js",
@@ -27,10 +27,10 @@
     "test": "echo no tests"
   },
   "dependencies": {
-    "@misk/common": "^0.0.28"
+    "@misk/common": "^0.0.29"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.15",
-    "@misk/webpack": "^0.0.16"
+    "@misk/dev": "^0.0.20",
+    "@misk/tslint": "^0.0.3"
   }
 }
