+++ misk/web/@misk/dev/package.json
@@ -1,9 +1,8 @@
 {
   "name": "@misk/dev",
-  "version": "0.0.16",
+  "version": "0.0.20",
   "description": "Microservice Kontainer Development Libraries",
   "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
-  "main": "tslint.base.json",
   "repository": {
     "type": "git",
     "url": "git@github.com:square/misk.git"
@@ -20,15 +19,23 @@
     "@types/react-router-redux": "^5.0.15",
     "@types/webpack": "^4.4.11",
     "@types/webpack-env": "^1.13.6",
-    "tslint": "^5.11.0",
-    "tslint-blueprint": "^0.1.0",
-    "tslint-clean-code": "^0.2.7",
-    "tslint-config-prettier": "^1.15.0",
-    "tslint-consistent-codestyle": "^1.13.3",
-    "tslint-eslint-rules": "^5.4.0",
-    "tslint-immutable": "^4.7.0",
-    "tslint-react": "^3.6.0",
-    "tslint-sonarts": "^1.7.0",
-    "typescript": "^3.0.3"
+    "awesome-typescript-loader": "^5.2.0",
+    "copy-webpack-plugin": "^4.5.2",
+    "cross-env": "^5.2.0",
+    "css-loader": "^1.0.0",
+    "file-loader": "^2.0.0",
+    "html-webpack-plugin": "^3.2.0",
+    "node-sass": "^4.9.3",
+    "sass-loader": "^7.1.0",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^0.23.0",
+    "typescript": "^3.0.3",
+    "webpack": "^4.17.2",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.7",
+    "webpack-merge": "^4.1.4"
+  },
+  "devDependencies": {
+    "@misk/common": "^0.0.28"
   }
 }
