--- misk/web/@misk/webpack/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@misk/webpack",
-  "version": "0.0.16",
-  "description": "Microservice Kontainer Webpack Build Configs",
-  "author": "Square/Misk Authors (https://github.com/square/misk/graphs/contributors)",
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:square/misk.git"
-  },
-  "scripts": {
-    "build:lib": "npm run lib",
-    "build": "npm run lib",
-    "dist": "npm run lib",
-    "lib": "echo no build",
-    "lint": "echo no lint",
-    "prepare": "yarn run lint && yarn run test && yarn run lib",
-    "reinstall": "rm -rf node_modules/ && yarn install",
-    "test": "echo no tests"
-  },
-  "dependencies": {
-    "awesome-typescript-loader": "^5.2.0",
-    "copy-webpack-plugin": "^4.5.2",
-    "cross-env": "^5.2.0",
-    "css-loader": "^1.0.0",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
-    "node-sass": "^4.9.3",
-    "sass-loader": "^7.1.0",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.22.1",
-    "webpack": "^4.16.5",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.5",
-    "webpack-merge": "^4.1.4"
-  },
-  "devDependencies": {
-    "@misk/common": "^0.0.27"
-  }
-}
