+++ misk/web/tabs/config/package.json
@@ -12,16 +12,16 @@
     "test": "jest --coverage"
   },
   "dependencies": {
-    "@misk/common": "^0.0.28",
-    "@misk/components": "^0.0.18",
+    "@misk/common": "^0.0.29",
+    "@misk/components": "^0.0.19",
     "@misk/tabs": "^0.0.1",
     "gray-matter": "^4.0.1",
     "js-yaml": "^3.12.0",
     "json2yaml": "^1.1.0"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.15",
-    "@misk/webpack": "^0.0.16"
+    "@misk/dev": "^0.0.20",
+    "@misk/tslint": "^0.0.3"
   },
   "miskTabWebpack": {
     "name": "Config",
