+++ misk/web/tabs/config/webpack.config.js
@@ -1,4 +1,4 @@
-const { MiskWebpackConfigBase } = require("@misk/webpack")
+const { MiskWebpackConfigBase } = require("@misk/dev")
 const path = require('path')
 const miskTabWebpack = require(path.join(process.cwd(), "package.json")).miskTabWebpack
 module.exports = MiskWebpackConfigBase(process.env.NODE_ENV, {
