+++ misk/web/tabs/loader/package.json
@@ -12,13 +12,13 @@
     "test": "jest --coverage"
   },
   "dependencies": {
-    "@misk/common": "^0.0.28",
-    "@misk/components": "^0.0.18",
+    "@misk/common": "^0.0.29",
+    "@misk/components": "^0.0.19",
     "@misk/tabs": "^0.0.1"
   },
   "devDependencies": {
-    "@misk/dev": "^0.0.15",
-    "@misk/webpack": "^0.0.16"
+    "@misk/dev": "^0.0.20",
+    "@misk/tslint": "^0.0.3"
   },
   "miskTabWebpack": {
     "name": "Loader",
